Update the authentication viewmodel
